Added validation to the delete method of the transaction service in case transaction does not exist, then it returns true

+++ src/main/java/com/kyoshisakuda/budget/account/transaction/TransactionServiceImpl.java
@@ -38,6 +38,8 @@ public class TransactionServiceImpl implements TransactionService {
     }
 
     public boolean deleteTransaction(long id) {
+        if (!repository.findById(id).isPresent())
+            return true;
         repository.deleteById(id);
         return (!repository.findById(id).isPresent());
     }
